requirements.txt
# LastPerson07Bot Production Dependencies - Updated 2025
python-telegram-bot>=20.7,<21.0
pymongo>=4.6.0,<5.0
motor>=3.3.0,<4.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
pydantic>=2.5.0,<3.0.0
opencv-python-headless>=4.8.0,<5.0.0
Pillow>=10.2.0,<11.0.0
requests>=2.32.0,<3.0.0
aiohttp>=3.9.0,<4.0.0
aiofiles>=23.0.0,<24.0.0
psutil>=5.9.0,<6.0.0
loguru>=0.7.0,<1.0.0
cryptography>=41.0.0,<43.0.0
aiolimiter>=1.1.0,<2.0.0
beautifulsoup4>=4.12.0,<5.0.0
markdown>=3.5.0,<4.0.0

# Optional for enhanced features
scrapy>=2.11.0,<3.0.0
numpy>=1.24.0,<2.0.0
scikit-image>=0.19.0,<1.0.0
face-recognition>=1.3.0,<2.0.0
